requirements.txt
alibabacloud-ecs20140526==4.0.3
allure-pytest==2.13.3
pytest==7.0.1
pytest-parallel==0.1.1
pytest_multithreading_allure==1.0.8
PyYAML==6.0.1
six==1.16.0
alibabacloud-darabonba-number==0.0.4
paramiko==3.4.0
scp==0.14.5
Flask==2.0.3
Werkzeug==2.0.3
influxdb==5.3.2
pymysql==1.0.2
peewee==3.17.3
redis==4.3.5

